Drop fbr/sb/sbm (fuzzy git branch switching); prefer git switch
Vanilla `git switch` handles tab completion, and if `fzf` is present and
configured properly, allows fuzzy matching vs branch names.  The only
thing it *doesn't* allow (and which these scripts did) is immediate
fuzzy matching on partial names (e.g. if you have `feature/1134` you
could do `fbr 1134` and if it would find the branch and switch to
it). However, `git switch` with `fzf` can effectively do that provided
you hit tab _before_ entering the branch name. Furthermore, there was a
bug here was `fzf` was picking filenames not branch names if you tried
to use it with tab completion - so we drop these flawed scripts in
favour of `git switch`.